Fix Login
It was annoying to have to click links to random places to login and do
anything, so make as much library as possible in order to share, then
add some meat to the root page.  Finally make projects view more updaty.
