+++ charts/ui/Chart.yaml
@@ -4,7 +4,7 @@ description: A Helm chart for deploying Unikorn UI
 
 type: application
 
-version: v0.1.9
-appVersion: v0.1.9
+version: v0.1.10
+appVersion: v0.1.10
 
 icon: https://assets.unikorn-cloud.org/assets/images/logos/dark-on-light/icon.png
